+++ Changelog.md
@@ -6,6 +6,9 @@ The changes for each individual (beta) release can be seen in each [release](htt
 
 - Make profile change value more robust
 - Fixed an issue where the target would not be correctly updated if it was not the first target
+- Fix an issue with the focuser move endpoint not being cancellable
+- Fix an issue where the mount info endpoint would fail
+- Add a new endpoint to get all equipment info bundled up `/equipment/info`
 
 ## 2.2.14.5